refactor(Controller): 뷰를 반환하는 컨트롤러와 json을 반환하는 컨트롤러를 분리
